dsfdsfds

+++ SafeUp/Models/SafeUpCollections/Users.cs
@@ -30,7 +30,7 @@ namespace SafeUp.Models.SafeUpCollections
         public override void AddRow(User detailRowModel)
         {
             InsertQuery = string.Format("insert into \"User\" values (default,'{0}','{1}','{2}','{3}'", detailRowModel.Login,
-                detailRowModel.Password, detailRowModel.UsedSpace, detailRowModel.AccountType);
+                detailRowModel.Password, detailRowModel.UsedSpace, (int)detailRowModel.AccountType);
             PostgreClient.SetData(InsertQuery);
 
         }
